+++ app/components/DonateTrees/components/paymentMethods/CreditCardForm.tsx
@@ -1,15 +1,7 @@
-import React, { Component } from 'react';
-import {
-  Text,
-  View,
-  ScrollView,
-  TouchableOpacity,
-  Image,
-  Platform,
-  TextInput
-} from 'react-native';
-import Icon from 'react-native-vector-icons/FontAwesome5';
-import styles from '../../../styles/donation/donation.native';
+import React, { Component } from "react";
+import { Text, View, TouchableOpacity, Image } from "react-native";
+import Icon from "react-native-vector-icons/FontAwesome5";
+import styles from "../../../../styles/donation/donation.native";
 import {
   visa,
   masterCard,
@@ -17,37 +9,14 @@ import {
   discover,
   creditCard,
   creditCardBack
-} from './../../../assets';
-import CheckBox from 'react-native-check-box';
-import {
-  CreditCardInput,
-  LiteCreditCardInput
-} from 'react-native-credit-card-input';
-
-import { TextField } from 'react-native-material-textfield';
-import i18n from '../../../locales/i18n';
+} from "../../../../assets";
+import CheckBox from "react-native-check-box";
+import { CreditCardInput } from "react-native-credit-card-input";
 
 export default class CreditCardForm extends Component {
   state = {
     creditCardInfo: false,
-    creditCardsList: [
-      // {
-      //   id: 0,
-      //   selected: false,
-      //   image: discover,
-      //   cardNumber: '*** 5997',
-      //   expiryDate: 'Expired',
-      //   cvv: '453'
-      // },
-      // {
-      //   id: 1,
-      //   selected: true,
-      //   image: americanExpress,
-      //   cardNumber: '*** 5997',
-      //   expiryDate: '10/20',
-      //   cvv: '234'
-      // }
-    ],
+    creditCardsList: [],
     showNewCreditCard: false,
     showPay: true,
     cardValues: null,
@@ -89,15 +58,14 @@ export default class CreditCardForm extends Component {
       cardValues: form,
       changeInCardValue: true
     });
-  }
-
-
-
-
+  };
 
   render() {
-
-    if (this.state.cardValues && this.state.cardValues.valid && this.state.changeInCardValue) {
+    if (
+      this.state.cardValues &&
+      this.state.cardValues.valid &&
+      this.state.changeInCardValue
+    ) {
       this.setState({
         changeInCardValue: false
       });
@@ -108,8 +76,8 @@ export default class CreditCardForm extends Component {
         cvc: this.state.cardValues.values.cvc,
         expMonth: expArr[0],
         expYear: expArr[1]
-      }
-      this.props.setcardValues(cardDetails)
+      };
+      this.props.setcardValues(cardDetails);
     }
     return (
       <View>
@@ -134,20 +102,20 @@ export default class CreditCardForm extends Component {
                 style={{ marginLeft: 10 }}
               />
             ) : (
-                <Icon
-                  name="chevron-down"
-                  size={14}
-                  color="rgba(0, 0, 0, 0.38)"
-                  style={{ marginLeft: 10 }}
-                />
-              )}
+              <Icon
+                name="chevron-down"
+                size={14}
+                color="rgba(0, 0, 0, 0.38)"
+                style={{ marginLeft: 10 }}
+              />
+            )}
           </TouchableOpacity>
           {/* Hidden until expanded by User */}
           {this.state.creditCardInfo ? (
             <View>
               {this.state.creditCardsList.map(creditCardsList => (
                 <TouchableOpacity
-                  style={{ flexDirection: 'row', marginTop: 28 }}
+                  style={{ flexDirection: "row", marginTop: 28 }}
                   onPress={() => {
                     this.selectCreditCard(creditCardsList.id);
                   }}
@@ -171,14 +139,14 @@ export default class CreditCardForm extends Component {
                     source={creditCardsList.image}
                     style={[styles.creditCardsDesign, { marginRight: 20 }]}
                   />
-                  <Text style={{ fontWeight: 'bold', marginRight: 10 }}>
+                  <Text style={{ fontWeight: "bold", marginRight: 10 }}>
                     {creditCardsList.cardNumber}
                   </Text>
 
                   <Text
                     style={
-                      creditCardsList.expiryDate == 'Expired'
-                        ? { color: 'red' }
+                      creditCardsList.expiryDate == "Expired"
+                        ? { color: "red" }
                         : null
                     }
                   >
@@ -190,7 +158,7 @@ export default class CreditCardForm extends Component {
               {/* View for New Credit Card  */}
               <View>
                 <TouchableOpacity
-                  style={{ flexDirection: 'row', marginTop: 28 }}
+                  style={{ flexDirection: "row", marginTop: 28 }}
                   onPress={() => {
                     this.selectNewCreditCard();
                   }}
@@ -213,11 +181,11 @@ export default class CreditCardForm extends Component {
                   <Text
                     style={{
                       fontSize: 14,
-                      fontFamily: 'OpenSans-Bold',
+                      fontFamily: "OpenSans-Bold",
                       lineHeight: 19,
                       letterSpacing: 0,
-                      textAlign: 'left',
-                      color: 'rgba(0, 0, 0, 0.6)'
+                      textAlign: "left",
+                      color: "rgba(0, 0, 0, 0.6)"
                     }}
                   >
                     Add New Payment method
@@ -227,17 +195,17 @@ export default class CreditCardForm extends Component {
                   <View style={{ marginTop: 18 }}>
                     <CreditCardInput
                       onChange={this._onChange}
-                      inputContainerStyle={{ maxWidth: 120, fontFamily: 'OpenSans-Regular' }}
+                      inputContainerStyle={{
+                        maxWidth: 120,
+                        fontFamily: "OpenSans-Regular"
+                      }}
                       cardImageFront={creditCard}
                       cardImageBack={creditCardBack}
-                      cardFontFamily={'OpenSans-Regular'}
-                      labelStyle={{ fontFamily: 'OpenSans-Regular' }}
+                      cardFontFamily={"OpenSans-Regular"}
+                      labelStyle={{ fontFamily: "OpenSans-Regular" }}
                       allowScroll
-                      placeholderColor={'#d5d5d5'}
+                      placeholderColor={"#d5d5d5"}
                     />
-                    {/* <LiteCreditCardInput onChange={this._onChange}
-                    // inputContainerStyle={{ maxWidth: 100 }}
-                    /> */}
 
                     <CheckBox
                       style={{ flex: 1, marginTop: 20 }}
@@ -247,11 +215,11 @@ export default class CreditCardForm extends Component {
                         });
                       }}
                       rightText={
-                        'Save this payment method in my account for future donations'
+                        "Save this payment method in my account for future donations"
                       }
                       rightTextStyle={styles.checkBoxText}
                       isChecked={this.state.isChecked}
-                      checkBoxColor={'#89b53a'}
+                      checkBoxColor={"#89b53a"}
                     />
                   </View>
                 ) : null}
@@ -265,25 +233,4 @@ export default class CreditCardForm extends Component {
       </View>
     );
   }
-}
-
-// Changes to be made in library page - CreditCardInput
-{
-  /* <ScrollView ref="Form"
-    keyboardShouldPersistTaps="always"
-    scrollEnabled={allowScroll}
-    showsHorizontalScrollIndicator={false}
-    style={s.form}> removed horizontal prop
-    <CCInput {...this._inputProps("number")}
-      keyboardType="numeric"
-      containerStyle={[s.inputContainer, inputContainerStyle, { width: CARD_NUMBER_INPUT_WIDTH }]} />
-    <View style={{ flexDirection: "row", marginTop: 21 }}> Added View
-      <CCInput {...this._inputProps("expiry")}
-        keyboardType="numeric"
-        containerStyle={[s.inputContainer, inputContainerStyle, { width: EXPIRY_INPUT_WIDTH }]} />
-      {requiresCVC &&
-        <CCInput {...this._inputProps("cvc")}
-          keyboardType="numeric"
-          containerStyle={[s.inputContainer, inputContainerStyle, { width: CVC_INPUT_WIDTH }]} />}
-    </View> */
 }
