+++ app/components/DonateTrees/screens/PaymentDetails.native.js
@@ -17,11 +17,11 @@ import { nextArrowWhite, paypal, paypalLogo } from "../../../assets";
 import styles from "../../../styles/donation/donation.native";
 import colors from "../../../utils/constants";
 import { formatNumber } from "../../../utils/utils";
-import CreditCardForm from "./../components/CreditCardForm";
+import CreditCardForm from "../components/PaymentMethods/CreditCardForm";
 import SafeAreaView from "react-native-safe-area-view";
 import { Header } from "./../components/Header";
 import PaymentLoader from "../components/PaymentLoader";
-import { handleNativePayPress } from "./../components/paymentMethods/nativePay";
+import { handleNativePayPress } from "./../components/PaymentMethods/nativePay";
 import Axios from "axios";
 
 export default function DonationStep3(props) {
